+++ stockMate/src/main/java/com/stockm8/persistence/StockDAO.java
@@ -33,12 +33,5 @@ public interface StockDAO {
     // 카테고리 목록 조회
     public List<CategoryVO> selectAllCategories() throws Exception;
     
-    // 바코드로 재고 조회
-  	public List<StockVO> selectQuantityCheckByBarcode(int businessId, String barcode) throws Exception;
- 	
-   	// 바코드로 재고 감소
-  	public int updateIncreseStock(int businessId, String barcode) throws Exception;
- 	
-  	// 바코드로 재고 감소 실시간 조회
-   	public int selectStockByBarcode(int businessId, String barcode) throws Exception;
+
 }
