+++ stockMate/src/main/java/com/stockm8/persistence/StockDAOImpl.java
@@ -95,36 +95,7 @@ public class StockDAOImpl implements StockDAO {
 
  
   
-  @Override
-	public List<StockVO> selectQuantityCheckByBarcode(int businessId, String barcode) throws Exception {
-		logger.info("selectReceiving() 호출");
-		Map<String, Object> paramMap = new HashMap();
-		paramMap.put("businessId", businessId);	   
-		paramMap.put("barcode", barcode);
-		
-		return sqlSession.selectList(NAMESPACE + "selectQuantityCheckByBarcode", paramMap);
-	}
 
-	@Override
-	public int updateIncreseStock(int businessId, String barcode) throws Exception {
-		logger.info("updateDecreaseStock() 호출");
-	    // 매개변수 묶기
-	    Map<String, Object> params = new HashMap<>();
-	    params.put("businessId", businessId);
-	    params.put("barcode", barcode);
-
-	    // SQL 실행
-	    return sqlSession.update(NAMESPACE + "updateIncreseStock", params);
-	}
-
-	@Override
-	public int selectStockByBarcode(int businessId, String barcode) throws Exception {
-		logger.info("selectStockByBarcode() 호출");
-		Map<String, Object> params = new HashMap<>();
-	    params.put("businessId", businessId);
-	    params.put("barcode", barcode);
-		return sqlSession.selectOne(NAMESPACE + "selectStockByBarcode", params);
-	}
 
 }
